+++ src/components/icon/icon.tsx
@@ -17,6 +17,7 @@ import _startCase from 'lodash/startCase';
 // TS file (dev/docs) or the JS file (distributed), and it's more effort than worth
 // to generate & git track a TS module definition for each icon component
 import { icon as empty } from './assets/empty.js';
+import { enqueueStateChange } from '../../services/react';
 
 const typeToPathMap = {
   addDataApp: 'app_add_data',
@@ -438,6 +439,10 @@ export type EuiIconProps = CommonProps &
      * Its value should be one or more element IDs
      */
     'aria-labelledby'?: string;
+    /**
+     * Callback when the icon has been loaded & rendered
+     */
+    onIconLoad?: () => void;
   };
 
 interface State {
@@ -513,13 +518,23 @@ export class EuiIcon extends PureComponent<EuiIconProps, State> {
       // eslint-disable-next-line prefer-template
       './assets/' + typeToPathMap[iconType] + '.js'
     ).then(({ icon }) => {
-      if (this.isMounted) {
-        this.setState({
-          icon,
-          iconTitle: iconType,
-          isLoading: false,
-        });
-      }
+      enqueueStateChange(() => {
+        if (this.isMounted) {
+          this.setState(
+            {
+              icon,
+              iconTitle: iconType,
+              isLoading: false,
+            },
+            () => {
+              const { onIconLoad } = this.props;
+              if (onIconLoad) {
+                onIconLoad();
+              }
+            }
+          );
+        }
+      });
     });
   };
 
@@ -531,6 +546,7 @@ export class EuiIcon extends PureComponent<EuiIconProps, State> {
       className,
       tabIndex,
       title,
+      onIconLoad,
       ...rest
     } = this.props;